// Portal-RH_Frontend-Servidor/src/app/models/solicitacoes/licenca-sem-vencimento/licenca-sem-vencimento.model.ts
export interface LicencaSV {
  matricula: number;
  cpf: string;
  rg: string;
  nome: string;
  localTrabalho: string;
  cargo: string;
  cep: string;
  endereco: string;
  bairro: string;
  municipio: string;
  numeroEndereco: string;
  complementoEndereco: string;
  conteudo: string;
  contribuicao: boolean;
}